refactor: restructure events
