+++ pandemonium/client_payload.go
@@ -1,0 +1,5 @@
+package pandemonium
+
+const (
+	PingOp OpcodeType = "PING"
+)
